Reworked logger + basic rewards for combat + combat screen as pop up
Now, CombatLogger instance is created for each combat encounter, that way its gonna be possible for more than one combat encounter to be active. Added basic rewards for combat, which are awarded once the combat screen is closed. Also added combat screen which is going to have an option to view combat log.